+++ alembic/versions/20250910_add_meetup_proposer_confirmer.py
@@ -3,7 +3,7 @@ import sqlalchemy as sa
 
 # revision identifiers, used by Alembic.
 revision = 'a20250910_meetup_cols'
-down_revision = None
+down_revision = '20250101_create_initial_tables'
 branch_labels = None
 depends_on = None