+++ alembic/versions/20250913_profile_house_system.py
@@ -10,7 +10,7 @@ import sqlalchemy as sa
 
 # revision identifiers, used by Alembic.
 revision = '20250913_profile_house_system'
-down_revision = 'a250913_tz_availslot'
+down_revision = '20250101_create_initial_tables'
 branch_labels = None
 depends_on = None