+++ alembic/versions/20251006_match_comment_localization.py
@@ -14,7 +14,7 @@ import sqlalchemy as sa
 
 # revision identifiers, used by Alembic.
 revision: str = "20251006_match_comment_localization"
-down_revision: Union[str, Sequence[str], None] = "20250930_add_refresh_tokens"
+down_revision: Union[str, Sequence[str], None] = "20250101_create_initial_tables"
 branch_labels: Union[str, Sequence[str], None] = None
 depends_on: Union[str, Sequence[str], None] = None